refactor(container): move sub-canvas render into plugin
