+++ apps/demo-free-layout/src/form-components/index.ts
@@ -6,4 +6,3 @@ export * from './form-header';
 export * from './form-item';
 export * from './type-tag';
 export * from './properties-edit';
-export * from './sub-canvas';
