+++ apps/demo-free-layout/src/nodes/loop/loop-form-render.tsx
@@ -1,13 +1,14 @@
 import { FormRenderProps, FlowNodeJSON } from '@flowgram.ai/free-layout-editor';
+import { SubCanvasRender } from '@flowgram.ai/free-container-plugin';
 
-import { FormHeader, FormContent, FormInputs, FormOutputs, SubCanvas } from '../../form-components';
+import { FormHeader, FormContent, FormInputs, FormOutputs } from '../../form-components';
 
 export const LoopFormRender = ({ form }: FormRenderProps<FlowNodeJSON>) => (
   <>
     <FormHeader />
     <FormContent>
       <FormInputs />
-      <SubCanvas />
+      <SubCanvasRender />
       <FormOutputs />
     </FormContent>
   </>
